refactor(reasoning): WS-3 Phase 1 — move tool-parsing.ts to kernel/utils/
Closes Cycle 1 (act ↔ decide) of the kernel capability DAG refactor.
tool-parsing.ts is a pure substrate (no Effect, no state) — its location
under capabilities/act/ was creating a spurious decide → act import edge
via arbitrator.ts pulling FINAL_ANSWER_RE / extractFinalAnswer.

Move: packages/reasoning/src/kernel/capabilities/act/tool-parsing.ts
   →  packages/reasoning/src/kernel/utils/tool-parsing.ts

Import sites updated (5):
  - kernel/capabilities/decide/arbitrator.ts:12  (../act/ → ../../utils/)
  - kernel/capabilities/reason/think.ts:46       (../act/ → ../../utils/)
  - kernel/capabilities/act/tool-execution.ts:32 (./ → ../../utils/)
  - strategies/reactive.ts:28                    (backward-compat re-export path)
  - tests/strategies/kernel/utils/tool-utils.test.ts:6 (test import)

No semantic changes to tool-parsing.ts itself. All exports preserved.
strategies/reactive.ts re-export of evaluateTransform retained for
backward compat (consumer audit deferred to a separate dispatch).

Gates:
  - Build: 38/38 green
  - Tests: 5750 pass / 0 fail / 23 skip (5773 total)
  - Typecheck: unchanged from WS-1 baseline (1 pre-existing error in
    packages/testing/src/gate/scenarios/cf-24-... not introduced here;
    that file is outside this dispatch's authority and untouched)
  - Cycle 1 (decide → act) edge eliminated: arbitrator.ts no longer
    imports from ../act/

Refs: wiki/Planning/Implementation-Plans/2026-05-28-ws-3-kernel-capability-dag.md §Phase 1
Refs: wiki/Planning/Implementation-Plans/2026-05-28-canonical-refactor.md §4 RC-2
