+++ packages/reasoning/src/kernel/utils/tool-parsing.ts
